1.52 commit

+++ views/include/scope.createcategoria.php
@@ -64,7 +64,20 @@ function letrasynumeros(e){
             <td><?php echo ucwords($row["fil_nom"]); ?></td>
               <td>
                   <a class="btn-datagrid" href="ActualizarCategorias-<?php echo $row['fil_id']; ?>-<?php echo $_SESSION['emp']['id']; ?>"><i class="fas fa-pen-square icono-accion"></i>Actualizar</a>
-                  <a class="btn-datagrid" href="EliminarCategoria-<?php echo $row['fil_id']; ?>-<?php echo $_SESSION['emp']['id']; ?>"><i class="fas fa-trash-alt icono-accion"></i>Eliminar</a>
+                  <a class="btn-datagrid" href="EliminarCategoria-<?php echo $row['fil_id']; ?>-<?php echo $_SESSION['emp']['id']; ?>" onclick="return confirmar(this)"><i class="fas fa-trash-alt icono-accion"></i>Eliminar</a>
+                  <script type="text/javascript">
+                  function confirmar()
+                  {
+                      if (confirm("Eliminar marca?")==false)
+                      {
+                          return false;
+                      }
+                      else
+                      {
+                          return true;
+                      }
+                  }
+                  </script>
               </td>
           </tr>
         <?php } ?>
